Add file pagination

+++ src/config.rs
@@ -4,7 +4,7 @@
 pub const FILE: &'static str = "primes.bin";
 
 // Maximum number RAM that the primes are allowed to take up. In bytes.
-pub const MAX_MEM_USAGE: u64 = 1073741824;
+pub const MAX_MEM_USAGE: usize = 1073741824;
 
 // Number of cores on the computer
 pub const CORES: usize = 4;
